majuscule

+++ import_export/objmanager.cpp
@@ -7,13 +7,13 @@ void ObjManager::writeToObj(const std::string& name, const std::vector<Vector3D>
 
     obj << "#vertices:\n";
 
-    for(int i = 0 ; i < vertex.size(); i++) {
+    for(size_t i = 0 ; i < vertex.size(); i++) {
         obj << "v " << std::setprecision(4) << vertex[i].x << " " << vertex[i].y << " " << vertex[i].z << "\n";
     }
 
     obj << "#faces:\n";
 
-    for(int i = 0 ; i < face.size(); i += 3){
+    for(size_t i = 0 ; i < face.size(); i += 3){
         obj << "f " << face[i] + 1 << " " << face[i+1] + 1 << " "<< face[i+2] + 1 << "\n";
 
     }
